Factor out calculator page to it's own file
